+++ src/test/kotlin/study/racinggame/service/DrawViewTest.kt
@@ -2,7 +2,7 @@ package study.racinggame.service
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import study.racinggame.domain.CarResponse
+import study.racinggame.presentation.dto.CarResponse
 import study.racinggame.presentation.view.DrawView
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
